// src/app/model/EmployeeDetails.ts
export class EmployeeDetails{

    "userId": number=0;

    "username": string="";

    "password": string="";

    "organizationNumber": number=0;

    "employeeEmail": string="";

    "gender": string="";

    "dateOfBirth": string="";

    "contactNumber": number=0;

    "city": string="";

    "designation": string="";

    "panNumber": string="";

    "status": string="";

    "organizationName":string="";

}